Factor out PrivilegesTask from TransportPrivilegesAction
